Implement ToTokens as literals for arrays, tuples, slices, and Vec.
This is consistent with the existing impls for String and str
that produce string literal syntax.

Slices and Vec use array literal syntax, a `&` or `vec!` prefix
needs to be added explicitly.

Arrays and tuples have impls up to 32 and 12 members, respectively.
This uses the same macros to generate impls as `libcore` does internally.