Allow to interpolate with `#{some.expression}` in addition to #var
Often, I am in situation in which i have to create named variable just
to give it to quote. So this patch allows to add arbitrary expressions
within #{...}

Does this break compatibility?
If somebody had used `#{...}` in a quote block before, this would now break
as it was teken literaly before and it is now an interpolation.
But since `#{...}` is not valid rust syntax, I would think that it is really
unlikely someone used that in its code.