Allow inline expressions within quote!()
This allows quote!() to contain "inline expressions".
These are expressions which are evaluated, not quoted.
This simplifies many situations where a procedural macro
needs to generate complex token streams.

For example, this code:

  let foo = if x {
    quote!{ something ... }
  } else {
    quote!{ more stuff ... }
  };

  quote! {
    fn foo() -> i32 { 42 + #foo }
  }

can be simplified to this:

  quote!{
    fn foo() -> i32 {
      42 + #{
        if x {
          quote!{ something ... }
        } else {
          quote!{ more stuff ... }
        }
      }
    }
  }

This eliminates the need to bind temporary token streams to variables,
then to refer to those variables by name. Not only does this eliminate
the need for a temporary name, but it can allow code that builds code
to more closely resemble the structure of the code that is generating.
Localized decisions can stay in their proper scope, rather than needing
to be hoisted above a complex quote!{...} expression.