Exclude compile tests from the published package
During a dependency review I noticed that quote includes the output
of it's compile tests in the published package. These files are not
required for building quote as dependency and make it a bit harder
to review the code.

This commit introduces a `include` directive in the `Cargo.toml` file to
explicitly include only those files required to build `quote`. This
excludes the compile test artifacts and also reduces the size of the
published crate from 35 files, 132.9KiB (30.7KiB compressed) to
16 files, 123.8KiB (28.1KiB compressed) which results in a 76GB/Month
traffic reduction for crates.io assuming the current 31 million
downloads and the difference in the compressed package size.